chore (refs T22614): move statement lib folder to client/js.
