+++ client/js/bundles/core/listDraft.js
@@ -19,7 +19,7 @@ import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import { prefixClass } from 'demosplan-ui/lib'
 import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
-import StatementForm from '@DemosPlanStatementBundle/lib/StatementForm'
+import StatementForm from '@DpJs/lib/statement/StatementForm'
 
 const components = {
   DpMapModal,
