+++ client/js/bundles/map/publicDetail.js
@@ -25,7 +25,7 @@ import Map from '@DpJs/components/map/publicdetail/Map'
 import MapTools from '@DpJs/components/map/publicdetail/controls/MapTools'
 import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
 import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
-import StatementForm from '@DemosPlanStatementBundle/lib/StatementForm'
+import StatementForm from '@DpJs/lib/statement/StatementForm'
 import TableWrapper from '@DpJs/lib/TableWrapper'
 
 //  Vuex store modules (to be registered on core bundle vuex store)
