+++ client/js/bundles/statement/assessmentStatement.js
@@ -11,7 +11,7 @@
  * This is the entrypoint for assessment_statement.html.twig and cluster_detail.html.twig
  */
 
-import AssessmentStatement from '@DemosPlanStatementBundle/lib/AssessmentStatement'
+import AssessmentStatement from '@DpJs/lib/statement/AssessmentStatement'
 import AssessmentTableStore from '@DemosPlanStatementBundle/store/AssessmentTable'
 import DetailView from '@DpJs/components/statement/assessmentTable/DetailView/DetailView'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
