+++ client/js/lib/statement/AssessmentStatement.js
