+++ client/js/lib/statement/DeleteFragmentButton.js
